better sorting, ui improvements

+++ src/main/kotlin/lu/gbraun/geo_exporter/dto/OsmPolygonDto.kt
@@ -1,10 +1,9 @@
 package lu.gbraun.geo_exporter.dto
 
-import org.geojson.Feature
 import org.geojson.Polygon
 
 data class OsmPolygonDto(
-    var id: Int? = null,
+    var id: Long? = null,
     var name: String? = null,
     var adminLevel: Int? = null,
     var way: Polygon? = null,
